feat(web/types): add follow-up TypeScript types
- Create centralized types in src/web/src/types/followup.ts
  - FollowUpStatus enum (Pending, Contacted, NoResponse, Connected, Declined)
  - FollowUpDto, UpdateFollowUpStatusRequest, AssignFollowUpRequest interfaces
- Update features/followups/api.ts to import from types file
- Re-export types for backward compatibility

Closes #311
